added support for variable definition

+++ Parser/NumberNode.cs
@@ -6,11 +6,11 @@ namespace Interpreter
 	public class NumberNode : ASTNode
 	{
 
-		public float value { get; private set; }
+		public Double value { get; private set; }
 
 		public NumberNode(Token token, long start, long end, string value) : base(token.type, start, end)
 		{
-			this.value = float.Parse(value);
+			this.value = Double.Parse(value);
 		}
 
 		public override void accept(IInterpreter interpreter)
